requirements.txt
python-ffmpeg
deepfilternet
PySoundFile
av
openai-whisper
ffmpeg
torchvision
torchaudio==2.7.0